+++ noodles-gtf/examples/gtf_write.rs
@@ -10,7 +10,7 @@ fn main() -> io::Result<()> {
     let stdout = io::stdout().lock();
     let mut writer = gtf::io::Writer::new(stdout);
 
-    let record = gtf::Record::default();
+    let record = gtf::RecordBuf::default();
     writer.write_record(&record)?;
 
     Ok(())