+++ noodles-gtf/src/io/writer/line.rs
@@ -4,15 +4,15 @@ mod record;
 use std::io::{self, Write};
 
 use self::{comment::write_comment, record::write_record};
-use crate::Line;
+use crate::LineBuf;
 
-pub(super) fn write_line<W>(writer: &mut W, line: &Line) -> io::Result<()>
+pub(super) fn write_line<W>(writer: &mut W, line: &LineBuf) -> io::Result<()>
 where
     W: Write,
 {
     match line {
-        Line::Comment(s) => write_comment(writer, s)?,
-        Line::Record(record) => write_record(writer, record)?,
+        LineBuf::Comment(s) => write_comment(writer, s)?,
+        LineBuf::Record(record) => write_record(writer, record)?,
     }
 
     write_newline(writer)?;