+++ noodles-gtf/src/io/writer/line/record/attributes.rs
@@ -3,7 +3,7 @@ mod field;
 use std::io::{self, Write};
 
 use self::field::write_field;
-use crate::record::Attributes;
+use crate::record_buf::Attributes;
 
 pub(super) fn write_attributes<W>(writer: &mut W, attributes: &Attributes) -> io::Result<()>
 where
@@ -31,7 +31,7 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::record::attributes::Entry;
+    use crate::record_buf::attributes::Entry;
 
     #[test]
     fn test_write_attributes() -> io::Result<()> {